Removed OTOS sensor offset, added side enum, cleaned up code, and reformatted opmodes
I changed the boolean saying whether the robot was on blue alliance to an enum called AllianceSide. The otos sensor has had its offset removed because I realized it only tracks changes which means adding the offset to the getposition and setposition method was not needed ig. Also I refactored the packages to place opmodes in auto and teleop packages - which required adding protected instead of package private to their class methods